removed parametrized constructors from domain classes

+++ Core/src/com/mfino/domain/Base.java
@@ -16,7 +16,8 @@ public class Base {
 	public static final String FieldName_LastUpdateTime = "lastUpdateTime";
 	public static final String FieldName_CreateTime = "createTime";
 	public static final String FieldName_CreatedBy = "createdby";
-	protected BigDecimal id;
+	
+	protected Long id;
 	protected long version;
 	protected Timestamp lastupdatetime;
 	protected String updatedby;
@@ -25,11 +26,11 @@ public class Base {
 	
 	@Id
 	@Column(name = "ID", unique = true, nullable = false, scale = 0)
-	public BigDecimal getId() {
+	public Long getId() {
 		return this.id;
 	}
 
-	public void setId(BigDecimal id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
