+++ Core/src/com/mfino/domain/MfsDenominations.java
@@ -19,7 +19,7 @@ public class MfsDenominations extends Base implements java.io.Serializable {
 	private static final long serialVersionUID = 1L;
 	public static final String FieldName_MFSBillerPartnerByMFSID = "mfsid";
 	public static final String FieldName_DenominationAmount = "denominationamount";
-	private BigDecimal mfsid;
+	private Long mfsid;
 	private BigDecimal denominationamount;
 	private String description;
 	private String productcode;
@@ -27,9 +27,9 @@ public class MfsDenominations extends Base implements java.io.Serializable {
 	public MfsDenominations() {
 	}
 
-	public MfsDenominations(BigDecimal id, Timestamp lastupdatetime,
+	public MfsDenominations(Long id, Timestamp lastupdatetime,
 			String updatedby, Timestamp createtime, String createdby,
-			BigDecimal mfsid, BigDecimal denominationamount) {
+			Long mfsid, BigDecimal denominationamount) {
 		this.id = id;
 		this.lastupdatetime = lastupdatetime;
 		this.updatedby = updatedby;
@@ -39,9 +39,9 @@ public class MfsDenominations extends Base implements java.io.Serializable {
 		this.denominationamount = denominationamount;
 	}
 
-	public MfsDenominations(BigDecimal id, Timestamp lastupdatetime,
+	public MfsDenominations(Long id, Timestamp lastupdatetime,
 			String updatedby, Timestamp createtime, String createdby,
-			BigDecimal mfsid, BigDecimal denominationamount,
+			Long mfsid, BigDecimal denominationamount,
 			String description, String productcode) {
 		this.id = id;
 		this.lastupdatetime = lastupdatetime;
@@ -56,11 +56,11 @@ public class MfsDenominations extends Base implements java.io.Serializable {
 
 	
 	@Column(name = "MFSID", nullable = false, scale = 0)
-	public BigDecimal getMfsid() {
+	public Long getMfsid() {
 		return this.mfsid;
 	}
 
-	public void setMfsid(BigDecimal mfsid) {
+	public void setMfsid(Long mfsid) {
 		this.mfsid = mfsid;
 	}
